+++ lib/ruby_speech/grxml.rb
@@ -8,6 +8,7 @@ module RubySpeech
     GRXML_NAMESPACE = 'http://www.w3.org/2001/06/grammar'
 
     %w{
+      builtins
       grammar
       rule
       item